Fixed support for oracle as query backend
This commit modifies gemfile in order to allow complete build of docker
image when enabling oci and oracle-enhanced on gemfile (see
https://github.com/ankane/blazer-docker/issues/6 for details).
Additionally introduces a Dockefile to be used in this case because
oracle instaclient is not working with alpine's musl libc. This new
container is based on redhat ubi8 OS with ruby 2.7 runtime and installs
oracle instaclient packages directly from oracle website.